[fix #1]Authentication and Registration

+++ app/Http/Controllers/TransactionsController.php
@@ -4,12 +4,23 @@ namespace App\Http\Controllers;
 
 use App\Models\Transactions;
 use Illuminate\Http\Request;
+use Illuminate\Routing\Controllers\HasMiddleware;
+use Illuminate\Routing\Controllers\Middleware;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 use function PHPUnit\Framework\returnSelf;
 
-class TransactionsController extends Controller
+class TransactionsController extends Controller implements HasMiddleware
 {
+
+    public static function middleware()
+    {
+        return [
+            new Middleware('auth:sanctum',except: ['index','show'])
+        ];
+    }
+
     /**
      * Display a listing of the resource.
      */
@@ -31,29 +42,33 @@ class TransactionsController extends Controller
             'description' => 'nullable|string|max:255',
         ]);
 
-        $fields['user_id'] = Auth::id();
+        // $fields['user_id'] = Auth::id();
 
-        Transactions::create($fields);
+        // Transactions::create($fields);
 
-        return ['transactions' => $fields];
+        $post = $request->user()->transactions()->create($fields);
+
+        return ['transactions' => $post];
     }
 
     /**
      * Display the specified resource.
      */
-    public function show(Transactions $transactions)
+    public function show($id)
     {
+        $transactions = Transactions::findOrFail($id);
+
         return ['transaction' => $transactions];
     }
 
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, Transactions $transactions)
+    public function update(Request $request, $id)
     {
-        if ($transactions->user_id !== Auth::id()) {
-            return response()->json(['message' => 'Unauthorized access'], 403);
-        }
+        $transactions = Transactions::findOrFail($id);
+
+        Gate::authorize('modify', $transactions);
 
         $fields = $request->validate([
             'amount' => 'required|numeric',
@@ -64,14 +79,19 @@ class TransactionsController extends Controller
 
         $transactions->update($fields);
 
-        return response()->json(["message" => "transactions updated succesfully!"]);
+        return ["message" => "Transactions succesfully Updated",
+                "transactions" => $transactions];
     }
 
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Transactions $transactions)
+    public function destroy($id)
     {
+        $transactions = Transactions::findOrFail($id);
+
+        Gate::authorize('modify', $transactions);
+
         $transactions->delete();
         return response()->json(["message" => "transactions deleted succesfully"]);
     }
